@@ -203,7 +203,7 @@ abstract class Expression implements Spec {
       );
 
   /// Return `var {name} = {this}`.
-  Expression assignVar(String name, [Reference type]) => BinaryExpression._(
+  Expression assignVar(String name, [Reference? type]) => BinaryExpression._(
         type == null
             ? LiteralExpression._('var $name')
             : BinaryExpression._(
@@ -216,7 +216,7 @@ abstract class Expression implements Spec {
       );
 
   /// Return `final {name} = {this}`.
-  Expression assignFinal(String name, [Reference type]) => BinaryExpression._(
+  Expression assignFinal(String name, [Reference? type]) => BinaryExpression._(
         type == null
             ? const LiteralExpression._('final')
             : BinaryExpression._(
@@ -229,7 +229,7 @@ abstract class Expression implements Spec {
       );
 
   /// Return `const {name} = {this}`.
-  Expression assignConst(String name, [Reference type]) => BinaryExpression._(
+  Expression assignConst(String name, [Reference? type]) => BinaryExpression._(
         type == null
             ? const LiteralExpression._('const')
             : BinaryExpression._(
@@ -313,7 +313,7 @@ class ToCodeExpression implements Code {
   const ToCodeExpression(this.code, [this.isStatement = false]);
 
   @override
-  R accept<R>(CodeVisitor<R> visitor, [R context]) =>
+  R accept<R>(CodeVisitor<R> visitor, [R? context]) =>
       (visitor as ExpressionVisitor<R>).visitToCodeExpression(this, context);
 
   @override
@@ -324,15 +324,15 @@ class ToCodeExpression implements Code {
 ///
 /// **INTERNAL ONLY**.
 abstract class ExpressionVisitor<T> implements SpecVisitor<T> {
-  T visitToCodeExpression(ToCodeExpression code, [T context]);
-  T visitBinaryExpression(BinaryExpression expression, [T context]);
-  T visitClosureExpression(ClosureExpression expression, [T context]);
-  T visitCodeExpression(CodeExpression expression, [T context]);
-  T visitInvokeExpression(InvokeExpression expression, [T context]);
-  T visitLiteralExpression(LiteralExpression expression, [T context]);
-  T visitLiteralListExpression(LiteralListExpression expression, [T context]);
-  T visitLiteralSetExpression(LiteralSetExpression expression, [T context]);
-  T visitLiteralMapExpression(LiteralMapExpression expression, [T context]);
+  T visitToCodeExpression(ToCodeExpression code, [T? context]);
+  T visitBinaryExpression(BinaryExpression expression, [T? context]);
+  T visitClosureExpression(ClosureExpression expression, [T? context]);
+  T visitCodeExpression(CodeExpression expression, [T? context]);
+  T visitInvokeExpression(InvokeExpression expression, [T? context]);
+  T visitLiteralExpression(LiteralExpression expression, [T? context]);
+  T visitLiteralListExpression(LiteralListExpression expression, [T? context]);
+  T visitLiteralSetExpression(LiteralSetExpression expression, [T? context]);
+  T visitLiteralMapExpression(LiteralMapExpression expression, [T? context]);
 }
 
 /// Knowledge of how to write valid Dart code from [ExpressionVisitor].
@@ -341,7 +341,7 @@ abstract class ExpressionVisitor<T> implements SpecVisitor<T> {
 abstract class ExpressionEmitter implements ExpressionVisitor<StringSink> {
   @override
   StringSink visitToCodeExpression(ToCodeExpression expression,
-      [StringSink output]) {
+      [StringSink? output]) {
     output ??= StringBuffer();
     expression.code.accept(this, output);
     if (expression.isStatement) {
@@ -352,7 +352,7 @@ abstract class ExpressionEmitter implements ExpressionVisitor<StringSink> {
 
   @override
   StringSink visitBinaryExpression(BinaryExpression expression,
-      [StringSink output]) {
+      [StringSink? output]) {
     output ??= StringBuffer();
     expression.left.accept(this, output);
     if (expression.addSpace) {
@@ -370,14 +370,14 @@ abstract class ExpressionEmitter implements ExpressionVisitor<StringSink> {
 
   @override
   StringSink visitClosureExpression(ClosureExpression expression,
-      [StringSink output]) {
+      [StringSink? output]) {
     output ??= StringBuffer();
     return expression.method.accept(this, output);
   }
 
   @override
   StringSink visitCodeExpression(CodeExpression expression,
-      [StringSink output]) {
+      [StringSink? output]) {
     output ??= StringBuffer();
     final visitor = this as CodeVisitor<StringSink>;
     return expression.code.accept(visitor, output);
@@ -385,45 +385,45 @@ abstract class ExpressionEmitter implements ExpressionVisitor<StringSink> {
 
   @override
   StringSink visitInvokeExpression(InvokeExpression expression,
-      [StringSink output]) {
-    output ??= StringBuffer();
+      [StringSink? output]) {
+    final out = output ??= StringBuffer();
     return _writeConstExpression(
-        output, expression.type == InvokeExpressionType.constInstance, () {
-      expression.target.accept(this, output);
+        out, expression.type == InvokeExpressionType.constInstance, () {
+      expression.target.accept(this, out);
       if (expression.name != null) {
-        output..write('.')..write(expression.name);
+        out..write('.')..write(expression.name);
       }
       if (expression.typeArguments.isNotEmpty) {
-        output.write('<');
-        visitAll<Reference>(expression.typeArguments, output, (type) {
-          type.accept(this, output);
+        out.write('<');
+        visitAll<Reference>(expression.typeArguments, out, (type) {
+          type.accept(this, out);
         });
-        output.write('>');
+        out.write('>');
       }
-      output.write('(');
-      visitAll<Spec>(expression.positionalArguments, output, (spec) {
-        spec.accept(this, output);
+      out.write('(');
+      visitAll<Spec>(expression.positionalArguments, out, (spec) {
+        spec.accept(this, out);
       });
       if (expression.positionalArguments.isNotEmpty &&
           expression.namedArguments.isNotEmpty) {
-        output.write(', ');
+        out.write(', ');
       }
-      visitAll<String>(expression.namedArguments.keys, output, (name) {
-        output..write(name)..write(': ');
-        expression.namedArguments[name].accept(this, output);
+      visitAll<String>(expression.namedArguments.keys, out, (name) {
+        out..write(name)..write(': ');
+        expression.namedArguments[name]!.accept(this, out);
       });
-      return output..write(')');
+      return out..write(')');
     });
   }
 
   @override
   StringSink visitLiteralExpression(LiteralExpression expression,
-      [StringSink output]) {
+      [StringSink? output]) {
     output ??= StringBuffer();
     return output..write(expression.literal);
   }
 
-  void _acceptLiteral(Object literalOrSpec, StringSink output) {
+  void _acceptLiteral(Object? literalOrSpec, StringSink output) {
     if (literalOrSpec is Spec) {
       literalOrSpec.accept(this, output);
       return;
@@ -436,71 +436,71 @@ abstract class ExpressionEmitter implements ExpressionVisitor<StringSink> {
   @override
   StringSink visitLiteralListExpression(
     LiteralListExpression expression, [
-    StringSink output,
+    StringSink? output,
   ]) {
-    output ??= StringBuffer();
+    final out = output ??= StringBuffer();
 
     return _writeConstExpression(output, expression.isConst, () {
       if (expression.type != null) {
-        output.write('<');
-        expression.type.accept(this, output);
-        output.write('>');
+        out.write('<');
+        expression.type!.accept(this, output);
+        out.write('>');
       }
-      output.write('[');
-      visitAll<Object>(expression.values, output, (value) {
-        _acceptLiteral(value, output);
+      out.write('[');
+      visitAll<Object>(expression.values, out, (value) {
+        _acceptLiteral(value, out);
       });
-      return output..write(']');
+      return out..write(']');
     });
   }
 
   @override
   StringSink visitLiteralSetExpression(
     LiteralSetExpression expression, [
-    StringSink output,
+    StringSink? output,
   ]) {
-    output ??= StringBuffer();
+    final out = output ??= StringBuffer();
 
     return _writeConstExpression(output, expression.isConst, () {
       if (expression.type != null) {
-        output.write('<');
-        expression.type.accept(this, output);
-        output.write('>');
+        out.write('<');
+        expression.type!.accept(this, output);
+        out.write('>');
       }
-      output.write('{');
-      visitAll<Object>(expression.values, output, (value) {
-        _acceptLiteral(value, output);
+      out.write('{');
+      visitAll<Object>(expression.values, out, (value) {
+        _acceptLiteral(value, out);
       });
-      return output..write('}');
+      return out..write('}');
     });
   }
 
   @override
   StringSink visitLiteralMapExpression(
     LiteralMapExpression expression, [
-    StringSink output,
+    StringSink? output,
   ]) {
-    output ??= StringBuffer();
-    return _writeConstExpression(output, expression.isConst, () {
+    final out = output ??= StringBuffer();
+    return _writeConstExpression(out, expression.isConst, () {
       if (expression.keyType != null) {
-        output.write('<');
-        expression.keyType.accept(this, output);
-        output.write(', ');
+        out.write('<');
+        expression.keyType!.accept(this, out);
+        out.write(', ');
         if (expression.valueType == null) {
-          const Reference('dynamic', 'dart:core').accept(this, output);
+          const Reference('dynamic', 'dart:core').accept(this, out);
         } else {
-          expression.valueType.accept(this, output);
+          expression.valueType!.accept(this, out);
         }
-        output.write('>');
+        out.write('>');
       }
-      output.write('{');
-      visitAll<Object>(expression.values.keys, output, (key) {
+      out.write('{');
+      visitAll<Object>(expression.values.keys, out, (key) {
         final value = expression.values[key];
-        _acceptLiteral(key, output);
-        output.write(': ');
-        _acceptLiteral(value, output);
+        _acceptLiteral(key, out);
+        out.write(': ');
+        _acceptLiteral(value, out);
       });
-      return output..write('}');
+      return out..write('}');
     });
   }
 
